Add a `queued` field to the `Job` model
The goal is to be able to know if a job is currently managed by a queue,
regardless its real status: waiting, delayed or running.

With this we can call `add_job` and not creating a new job even if the
current existing one is in the `running` or `delayed` status, which
wasn't the case before this because with used to check the `status`
field.

Closes #8